SEM Graph with t test on cluster level

Perform SEM-based graphs with t-test on cluster level for Seurat objects. Calculates mean expression values and SEM for selected features, and visualizes them. Performs pairwise t-tests comparing conditions, with optional custom control condition and clustering. Optionally returns a summary data frame.

Usage

DO.BarplotClustert(
  Seu_object,
  Features,
  ListTest = NULL,
  returnValues = FALSE,
  ctrl.condition = NULL,
  group.by = "condition",
  returnPlot = FALSE
)

Arguments

Seu_object

Combined Seu_object

Features

Vector containing featurenames

ListTest

List for which conditions t-test will be performed, if NULL always against provided CTRL

returnValues

return df.melt.sum data frame containing means and SEM for the set group

ctrl.condition

set your ctrl condition, relevant if running with empty comparison List

group.by

select the seurat object slot where your conditions can be found, default conditon

Value

a ggplot or a dataframe

Author

Mariano Ruz Jurado

Examples

if (FALSE) { # \dontrun{

(optional)
ListTest <- list()
ListTest[[1]] <- c("CTRL", "CONDITION")

DO.BarplotClustert(
  Seu_object = Seurat,
  Features = "CDH5",
  ListTest = ListTest,
  ctrl.condition = "CTRL",
  group.by="condition"
)
} # }