DO.FullRecluster

Performs iterative reclustering on each major cluster found by FindClusters in a Seurat object. It refines the clusters using the FindSubCluster function for better resolution and fine-tuned annotation. The new clustering results are stored in a metadata column called annotation_recluster. Suitable for improving cluster precision and granularity after initial clustering.

Usage

DO.FullRecluster(
  Seu_object,
  over_clustering = "seurat_clusters",
  res = 0.5,
  algorithm = 4,
  graph.name = "RNA_snn"
)

Arguments

Seu_object

The seurat object

over_clustering

Column in metadata in object with clustering assignments for cells, default seurat_clusters

res

Resolution for the new clusters, default 0.5

algorithm

Set one of the available algorithms found in FindSubCLuster function, default = 4: leiden

Value

a Seurat Object with new clustering named annotation_recluster

Author

Mariano Ruz Jurado

Examples

if (FALSE) { # \dontrun{

DO.FullRecluster(
  Seu_object = Seurat
)
} # }