DO Celltypist

Runs the CellTypist model on a Seurat object to predict cell type labels, storing the results as metadata. If the number of cells is less than the specified threshold, it returns NAs for the labels. Optionally updates the CellTypist models and returns the probability matrix. Useful for annotating cell types in single-cell RNA sequencing datasets.

Usage

DO.CellTypist(
  Seu_object,
  modelName = "Healthy_Adult_Heart.pkl",
  minCellsToRun = 200,
  runCelltypistUpdate = TRUE,
  over_clustering = "seurat_clusters",
  assay_normalized = "RNA",
  returnProb = FALSE,
  SeuV5 = T
)

Arguments

Seu_object

The seurat object

minCellsToRun

If the input seurat object has fewer than this many cells, NAs will be added for all expected columns and celltypist will not be run.

runCelltypistUpdate

If true, –update-models will be run for celltypist prior to scoring cells.

over_clustering

Column in metadata in object with clustering assignments for cells, default seurat_clusters

assay_normalized

Assay with log1p normalized expressions

returnProb

will additionally return the probability matrix, return will give a list with the first element beeing the object and second prob matrix

Value

a seurat

Author

Mariano Ruz Jurado

Examples

if (FALSE) { # \dontrun{

DO.CellTypist(
  Seu_object = Seurat,
  modelName = "Healthy_Adult_Heart.pkl",
  runCelltypistUpdate = TRUE,
  over_clustering = "seurat_clusters",
  SeuV5=T
)
} # }