DO.UMAP

Creates a polished UMAP plot using Seurat's DimPlot or FeaturePlot functions. It allows customization of colors, labels, and other plot elements for better visualization. The function handles both cluster-based visualizations and gene-based visualizations in a UMAP plot. Ideal for refining UMAP outputs with added flexibility and enhanced presentation.

Usage

DO.UMAP(
  Seu_object,
  FeaturePlot = F,
  features = NULL,
  group.by = "seurat_clusters",
  umap_colors = NULL,
  text_size = 14,
  label = T,
  order = T,
  plot.title = T,
  legend.position = "none",
  ...
)

Arguments

Seu_object

The seurat object

FeaturePlot

Is it going to be a Dimplot or a FeaturePlot?

features

features for Featureplot

group.by

grouping of plot in DImplot and defines in featureplot the labels

...

Further arguments passed to DimPlot or FeaturePlot function from Seurat

Value

Plot with Refined colors and axes

Author

Mariano Ruz Jurado

Examples

if (FALSE) { # \dontrun{

DO.UMAP(
  Seu_object = Seurat,
  group.by="seurat_clusters"
)

DO.UMAP(
  Seu_object = Seurat,
  FeaturePlot=T,
  features=c("CDH5","TTN")
)
} # }