DOtools: A Toolkit for scRNA Data Analysis

The `DOtools` package provides a set of functions for advanced data processing, visualisation, and statistical analysis in Seurat objects. It includes functions for cell-type prediction, reclustering, creating polished UMAP plots, subsetting Seurat objects, and various statistical analyses like Wilcoxon tests and SEM graphs.

Details

This package includes the following functions:

DO.BoxPlot: A function for creating box plots with Wilcoxon test results.

DO.CellTypist: A function for running CellTypist on Seurat objects to predict cell types.

DO.DietSeurat: A function for diet-based analysis of Seurat objects.

DO.Dotplot: A function for creating dot plots for visualizing gene expression.

DO.FullRecluster: A function for fine-grained reclustering of Seurat objects.

DO.BarplotClustert: A function for generating mean and SEM graphs for cluster-based analysis with t-tests.

DO.BarplotWilcox: A function for generating mean and SEM graphs with Wilcoxon test results.

DO.Subset: A function for subsetting Seurat objects based on metadata.

DO.UMAP: A function for creating polished UMAP plots using either DimPlot or FeaturePlot.

DO.VlnPlot: A function for generating violin plots with Wilcoxon test results.

DO.CellComposition: A function for visualizing and statistically analyzing cell-type composition changes across conditions using the Scanpro Python package, with support for bootstrapping, proportion plots, and customizable output.

DO.Import: A function for building a merged Seurat object from 10x software output, or directly from provided tables.

DO.CellBender: A function for running CellBender in a virtual conda env with provided raw count h5 files.

DO.SplitBarGSEA: A function for viusalizing GSEA result from a provided df from e.g. metascape

DO.scVI: A function for running the scVI Integration implemented in scvi-tools.

DO.TransferLabel: A function for transfering annotation from a subseted object to the original seurat object.

DO.PyEnv: A function for creating a conda envrionment holding all python packages needed for some functions.

DO.Correlation: A function for creating a correlation plot between provided samples in the category specified.

DO.Heatmap: A function for generating Heat maps on gene expression data.

DO.MultiDGE: A function for calculating DEGs on a single cell and speudo bulk level.

dot-Do.BarcodeRanks: A function for estimating the number of expected cells and droplets.

dot-QC.Vlnplot: A function for estimating the number of expected cells and droplets.

Author

Mariano Ruz Jurado, David Rodriguez Morales

See also